little_helper/requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
celery==5.5.3
redis==5.2.1
django-filter==25.1
pytest==8.4.1
pytest-django==4.11.1
factory_boy==3.3.3
Faker==37.5.3 